requirements.txt
# Core dependencies
python-telegram-bot[webhooks]==20.7
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Google Calendar API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Async and scheduling
apscheduler==3.10.4
aiohttp==3.9.1

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utils
pytz==2023.3
python-dateutil==2.8.2
psutil==5.9.6

# Diagnostic System dependencies
schedule==1.2.0
click==8.1.7
requests==2.31.0
networkx==3.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
watchdog==3.0.0